improvements to webthing
